Migrate the responsability of parsing response to the Call class
The Azure Storage Blob API returns responses in various formats and in
different parts of the HTTP response (sometimes in the content,
sometimes in the headers).
At this stage, it seems reasonable to store the state of in which format
a given API call returns the result and in which part of the HTTP
response in the Call class, as it gathers all the specific knowledge
around an API call.