Refactor the logic responsible of building the HTTP request
It does no longer, first, build the request object and, then, mutate its
state by setting the value of its headers.
Instead, the headers object is built and then passed during the creation
of the request object.

The intent of this change is to accomodate body content during request
creation and, also, tackle the need to set the
Content-Length header.